Reconfigure to use node-fetch
I ran into issues when trying to launch chromium, maybe because I'm
running from WSL? But then I wondered why we were launching an entire
browser to feth a webpage. Sure enough, the page is static and we can
just fetch the HTML.

I also had issues with ESModules vs CommonJS and the differing import
syntax. This combination seemed to work.

Also also added some stuff to gitignore that should probably be there.

This is a bad commit that combines lots of things. Sorry...